Completed InventoryManager component with kernel and enhanced interfaces
