+++ src/InventoryManagerKernel.java
@@ -1,8 +1,11 @@
+import components.standard.Standard;
+
 /**
  * Kernel interface for InventoryManager. Defines core operations of an
  * inventory system.
  */
-public interface InventoryManagerKernel {
+public interface InventoryManagerKernel extends Standard<InventoryManager> {
+
     /**
      * Adds an item to the inventory with the given quantity.
      *
@@ -10,6 +13,8 @@ public interface InventoryManagerKernel {
      *            Name of the item
      * @param quantity
      *            Quantity to be added
+     * @requires itemName != null && quantity > 0
+     * @ensures this.inventory[itemName] = #this.inventory[itemName] + quantity
      */
     void addItem(String itemName, int quantity);
 
@@ -20,6 +25,9 @@ public interface InventoryManagerKernel {
      *            Name of the item
      * @param quantity
      *            Quantity to be removed
+     * @requires itemName != null && quantity > 0 && this.inventory[itemName] >=
+     *           quantity
+     * @ensures this.inventory[itemName] = #this.inventory[itemName] - quantity
      */
     void removeItem(String itemName, int quantity);
 
@@ -29,6 +37,8 @@ public interface InventoryManagerKernel {
      * @param itemName
      *            Name of the item
      * @return true if the item exists, false otherwise
+     * @requires itemName != null
+     * @ensures \result = (this.inventory contains itemName)
      */
     boolean hasItem(String itemName);
 
@@ -38,6 +48,8 @@ public interface InventoryManagerKernel {
      * @param itemName
      *            Name of the item
      * @return Quantity of the item, or 0 if not found
+     * @requires itemName != null
+     * @ensures \result = this.inventory[itemName] if itemName exists, else 0
      */
     int getItemCount(String itemName);
 }
